fix: incorrect generation of nested serializers
Fixed serializer generation to include all nested serializers even if they are not
directly used. This ensures correct serialization of complex types like
Map<UInt, Map<UInt, UInt>>.

Previously, intermediate serializers were not generated, for example:
- MapOfUIntUInt (for Map<UInt, UInt>)

With this fix, all necessary serializers are now properly generated,
enabling correct serialization of nested types.

+++ example/src/main/kotlin/dev/s7a/example/ExamplePlugin.kt
@@ -44,7 +44,7 @@ class ExamplePlugin : JavaPlugin() {
                 list2 = List(5) { List(3) { Random.nextInt().toUInt() } },
                 set = List(5) { Random.nextInt().toUInt() }.toSet(),
                 arrayDeque = ArrayDeque(List(3) { Random.nextInt().toUInt() }),
-                map = (0..4).associate { Random.nextInt().toUInt() to Random.nextInt().toUInt() },
+                map = (0..4).associate { Random.nextLong().toULong() to Random.nextLong().toULong() },
                 map2 =
                     (0..4).associate {
                         Random.nextInt().toUInt() to
